requirements.txt
numpy
pandas
matplotlib.pyplot 
datetime
yfinance
scipy.stats
seaborn
tabulate
statsmodels.api 
cvxpy